move evm mono decorator out of example chain
